Infer bindings from C++ specs
