+++ src/ifcwrap/binding_generator/specs/cpp/ifcgeom.hpp
@@ -24,6 +24,10 @@
 #include "function_item_evaluator.h"
 #include "geometry_serializer_plugin.h"
 #include "document_serializer_plugin.h"
+#include "GeometrySerializer.h"
+#include "JsonSerializer.h"
+#include "RocksDbSerializer.h"
+#include "XmlSerializer.h"
 #include "../svgfill/src/svgfill.h"
 
 #include <Eigen/Dense>
@@ -92,6 +96,229 @@ IFCAPI_HANDLE(taxonomy_revolve, ifcopenshell::geometry::taxonomy::revolve, share
 IFCAPI_HANDLE(taxonomy_sweep_along_curve, ifcopenshell::geometry::taxonomy::sweep_along_curve, shared_ptr, shared_ptr)
 IFCAPI_HANDLE(taxonomy_node, ifcopenshell::geometry::taxonomy::node, shared_ptr, shared_ptr)
 IFCAPI_HANDLE(taxonomy_boolean_result, ifcopenshell::geometry::taxonomy::boolean_result, shared_ptr, shared_ptr)
+
+// Native methods included in the stable C ABI.
+IFCAPI_DISCOVER_METHOD(triangulation, edges, edges)
+IFCAPI_DISCOVER_METHOD(triangulation, edges_item_ids, edges_item_ids)
+IFCAPI_DISCOVER_METHOD(triangulation, faces, faces)
+IFCAPI_DISCOVER_METHOD(triangulation, item_ids, item_ids)
+IFCAPI_DISCOVER_METHOD(triangulation, material_ids, material_ids)
+IFCAPI_DISCOVER_METHOD(triangulation, materials, materials)
+IFCAPI_DISCOVER_METHOD(triangulation, normals, normals)
+IFCAPI_DISCOVER_METHOD(triangulation, polyhedral_faces_with_holes, polyhedral_faces_with_holes)
+IFCAPI_DISCOVER_METHOD(triangulation, polyhedral_faces_without_holes, polyhedral_faces_without_holes)
+IFCAPI_DISCOVER_METHOD(triangulation, uvs, uvs)
+IFCAPI_DISCOVER_METHOD(triangulation, verts, verts)
+IFCAPI_DISCOVER_METHOD(iterator, bounds_max, bounds_max)
+IFCAPI_DISCOVER_METHOD(iterator, bounds_min, bounds_min)
+IFCAPI_DISCOVER_METHOD(iterator, compute_bounds, compute_bounds, bool)
+IFCAPI_DISCOVER_METHOD(iterator, create, create)
+IFCAPI_DISCOVER_METHOD(iterator, file, file)
+IFCAPI_DISCOVER_METHOD(iterator, get, get)
+IFCAPI_DISCOVER_METHOD(iterator, getLog, get_log)
+IFCAPI_DISCOVER_METHOD(iterator, get_native, get_native)
+IFCAPI_DISCOVER_METHOD(iterator, get_object, get_object, int)
+IFCAPI_DISCOVER_METHOD(iterator, get_task_items, get_task_items)
+IFCAPI_DISCOVER_METHOD(iterator, get_task_products, get_task_products)
+IFCAPI_DISCOVER_METHOD(iterator, had_error_processing_elements, had_error_processing_elements)
+IFCAPI_DISCOVER_METHOD(iterator, initialize, initialize)
+IFCAPI_DISCOVER_METHOD(iterator, next, next)
+IFCAPI_DISCOVER_METHOD(iterator, progress, progress)
+IFCAPI_DISCOVER_METHOD(iterator, unit_magnitude, unit_magnitude)
+IFCAPI_DISCOVER_METHOD(iterator, unit_name, unit_name)
+IFCAPI_DISCOVER_METHOD(brep_representation, as_compound, as_compound, bool)
+IFCAPI_DISCOVER_METHOD(brep_representation, calculate_projected_surface_area, calculate_projected_surface_area, const std::shared_ptr<ifcopenshell::geometry::taxonomy::matrix4>&, double&, double&, double&)
+IFCAPI_DISCOVER_METHOD(brep_representation, entity, entity)
+IFCAPI_DISCOVER_METHOD(brep_representation, id, id)
+IFCAPI_DISCOVER_METHOD(brep_representation, item, item, int)
+IFCAPI_DISCOVER_METHOD(brep_representation, item_id, item_id, int)
+IFCAPI_DISCOVER_METHOD(brep_representation, settings, settings)
+IFCAPI_DISCOVER_METHOD(brep_representation, calculate_surface_area, calculate_surface_area, double&)
+IFCAPI_DISCOVER_METHOD(brep_representation, calculate_volume, calculate_volume, double&)
+IFCAPI_DISCOVER_METHOD(brep_representation, size, size)
+IFCAPI_DISCOVER_METHOD(element, context, context)
+IFCAPI_DISCOVER_METHOD(element, guid, guid)
+IFCAPI_DISCOVER_METHOD(element, id, id)
+IFCAPI_DISCOVER_METHOD(element, name, name)
+IFCAPI_DISCOVER_METHOD(element, parent_id, parent_id)
+IFCAPI_DISCOVER_METHOD(element, parents, parents)
+IFCAPI_DISCOVER_METHOD(element, product, product)
+IFCAPI_DISCOVER_METHOD(element, transformation, transformation)
+IFCAPI_DISCOVER_METHOD(element, type, type)
+IFCAPI_DISCOVER_METHOD(element, unique_id, unique_id)
+IFCAPI_DISCOVER_METHOD(brep_element, calculate_projected_surface_area, calculate_projected_surface_area, double&, double&, double&)
+IFCAPI_DISCOVER_METHOD(brep_element, geometry, geometry)
+IFCAPI_DISCOVER_METHOD(triangulation_element, geometry, geometry)
+IFCAPI_DISCOVER_METHOD(serialized_element, geometry, geometry)
+IFCAPI_DISCOVER_METHOD(conversion_result_shape, is_manifold, is_manifold)
+IFCAPI_DISCOVER_METHOD(conversion_result_shape, num_edges, num_edges)
+IFCAPI_DISCOVER_METHOD(conversion_result_shape, num_faces, num_faces)
+IFCAPI_DISCOVER_METHOD(conversion_result_shape, num_vertices, num_vertices)
+IFCAPI_DISCOVER_METHOD(conversion_result_shape, surface_area_along_direction, surface_area_along_direction, double, const std::shared_ptr<ifcopenshell::geometry::taxonomy::matrix4>&, double&, double&, double&)
+IFCAPI_DISCOVER_METHOD(conversion_result_shape, surface_genus, surface_genus)
+IFCAPI_DISCOVER_METHOD(function_item_evaluator, evaluation_points, evaluation_points)
+IFCAPI_DISCOVER_METHOD(function_item_evaluator, evaluation_points, evaluation_points_range, double, double, IFCAPI_INT32(unsigned int))
+IFCAPI_DISCOVER_METHOD(function_item_evaluator, evaluate, evaluate)
+IFCAPI_DISCOVER_METHOD(function_item_evaluator, evaluate, evaluate_range, double, double, IFCAPI_INT32(unsigned int))
+IFCAPI_DISCOVER_METHOD(serialization, brep_data, brep_data)
+IFCAPI_DISCOVER_METHOD(serialization, surface_style_ids, surface_style_ids)
+IFCAPI_DISCOVER_METHOD(serialization, surface_styles, surface_styles)
+IFCAPI_DISCOVER_METHOD(geometry_serializer, geometry_settings, geometry_settings)
+IFCAPI_DISCOVER_METHOD(geometry_serializer, settings, settings)
+IFCAPI_DISCOVER_METHOD(geometry_serializer, write, write_triangulation_element, const IfcGeom::TriangulationElement*)
+IFCAPI_DISCOVER_METHOD(geometry_serializer, write, write_brep_element, const IfcGeom::BRepElement*)
+IFCAPI_DISCOVER_METHOD(geometry_serializer, finalize, finalize)
+IFCAPI_DISCOVER_METHOD(geometry_serializer, isTesselated, is_tesselated)
+IFCAPI_DISCOVER_METHOD(geometry_serializer, is_streaming, is_streaming)
+IFCAPI_DISCOVER_METHOD(geometry_serializer, read, read, ifcopenshell::file&, const std::string&, const std::string&, GeometrySerializer::read_type)
+IFCAPI_DISCOVER_METHOD(geometry_serializer, ready, ready)
+IFCAPI_DISCOVER_METHOD(geometry_serializer, setFile, set_file, ifcopenshell::file*)
+IFCAPI_DISCOVER_METHOD(geometry_serializer, setUnitNameAndMagnitude, set_unit_name_and_magnitude, const std::string&, float)
+IFCAPI_DISCOVER_METHOD(geometry_serializer, writeHeader, write_header)
+IFCAPI_DISCOVER_METHOD(taxonomy_style, has_specularity, has_specularity)
+IFCAPI_DISCOVER_METHOD(taxonomy_style, has_transparency, has_transparency)
+IFCAPI_DISCOVER_METHOD(opaque_number, to_double, to_double)
+IFCAPI_DISCOVER_METHOD(opaque_number, to_string, to_string)
+IFCAPI_DISCOVER_METHOD(serializer, finalize, finalize)
+IFCAPI_DISCOVER_METHOD(serializer, is_streaming, is_streaming)
+IFCAPI_DISCOVER_METHOD(serializer, ready, ready)
+IFCAPI_DISCOVER_METHOD(serializer, setFile, set_file, ifcopenshell::file*)
+IFCAPI_DISCOVER_METHOD(serializer, writeHeader, write_header)
+IFCAPI_DISCOVER_METHOD(settings, get_type, get_type, const std::string&)
+IFCAPI_DISCOVER_METHOD(settings, setting_names, setting_names)
+IFCAPI_DISCOVER_METHOD(serializer_settings, get_type, get_type, const std::string&)
+IFCAPI_DISCOVER_METHOD(serializer_settings, setting_names, setting_names)
+IFCAPI_DISCOVER_METHOD(buffer, get_value, get_value)
+IFCAPI_DISCOVER_METHOD(buffer, is_ready, is_ready)
+IFCAPI_DISCOVER_METHOD(taxonomy_item, hash, hash)
+IFCAPI_DISCOVER_METHOD(taxonomy_item, identity, identity)
+IFCAPI_DISCOVER_METHOD(taxonomy_item, kind, kind)
+IFCAPI_DISCOVER_METHOD(tree, enable_face_styles, enable_face_styles)
+IFCAPI_DISCOVER_METHOD(tree, enable_face_styles, set_enable_face_styles, bool)
+IFCAPI_DISCOVER_METHOD(tree, add_file, add_file, ifcopenshell::file&, const ifcopenshell::geometry::Settings&)
+IFCAPI_DISCOVER_METHOD(tree, add_file, add_iterator, IfcGeom::Iterator&)
+IFCAPI_DISCOVER_METHOD(tree, clash_clearance_many, clash_clearance_many, const std::vector<express::Base>&, const std::vector<express::Base>&, double, bool)
+IFCAPI_DISCOVER_METHOD(tree, clash_collision_many, clash_collision_many, const std::vector<express::Base>&, const std::vector<express::Base>&, bool)
+IFCAPI_DISCOVER_METHOD(tree, clash_intersection_many, clash_intersection_many, const std::vector<express::Base>&, const std::vector<express::Base>&, double, bool)
+IFCAPI_DISCOVER_METHOD(tree, distances, distances)
+IFCAPI_DISCOVER_METHOD(tree, is_manifold, is_manifold, const std::vector<int>&)
+IFCAPI_DISCOVER_METHOD(tree, protrusion_distances, protrusion_distances)
+IFCAPI_DISCOVER_METHOD(tree, styles, styles)
+IFCAPI_DISCOVER_METHOD(tree, uint8_to_b64, uint8_to_b64, const std::vector<unsigned char>&)
+
+// Structured relationships that are not encoded by C++ types.
+IFCAPI_DISCOVER_CONSTRUCTOR(serializer, XmlSerializer, create_xml_serializer, explicit, _, _)
+IFCAPI_CONSTRUCTOR_PARAM(create_xml_serializer, file, file, ifcopenshell::file*)
+IFCAPI_CONSTRUCTOR_PARAM(create_xml_serializer, xml_filename, filename, const std::string&)
+IFCAPI_DISCOVER_CONSTRUCTOR(tree, IfcGeom::tree, create_tree, explicit, IFOPSH_WITH_OPENCASCADE, _)
+IFCAPI_DISCOVER_CONSTRUCTOR(tree, IfcGeom::tree, create_tree_from_file, explicit, IFOPSH_WITH_OPENCASCADE, _)
+IFCAPI_CONSTRUCTOR_PARAM(create_tree_from_file, f, file, ifcopenshell::file&)
+IFCAPI_DISCOVER_CONSTRUCTOR(tree, IfcGeom::tree, create_tree_from_file_with_settings, explicit, IFOPSH_WITH_OPENCASCADE, _)
+IFCAPI_CONSTRUCTOR_PARAM(create_tree_from_file_with_settings, f, file, ifcopenshell::file&)
+IFCAPI_CONSTRUCTOR_PARAM(create_tree_from_file_with_settings, settings, settings, const ifcopenshell::geometry::Settings&)
+IFCAPI_DISCOVER_CONSTRUCTOR(tree, IfcGeom::tree, create_tree_from_iterator, explicit, IFOPSH_WITH_OPENCASCADE, _)
+IFCAPI_CONSTRUCTOR_PARAM(create_tree_from_iterator, it, iterator, IfcGeom::Iterator&)
+IFCAPI_DISCOVER_CONSTRUCTOR(serializer, JsonSerializer, create_json_serializer, explicit, WITH_GLTF, "JSON serializer requires GLTF support (nlohmann_json)")
+IFCAPI_CONSTRUCTOR_PARAM(create_json_serializer, file, file, ifcopenshell::file*)
+IFCAPI_CONSTRUCTOR_PARAM(create_json_serializer, filename, filename, const std::string&)
+IFCAPI_DISCOVER_CONSTRUCTOR(serializer, RocksDbSerializer, create_rocksdb_serializer_streaming, explicit, IFOPSH_WITH_ROCKSDB, _)
+IFCAPI_CONSTRUCTOR_PARAM(create_rocksdb_serializer_streaming, input_filename, input_filename, const std::string&)
+IFCAPI_CONSTRUCTOR_PARAM(create_rocksdb_serializer_streaming, rocksdb_filename, rocksdb_filename, const std::string&)
+IFCAPI_DISCOVER_CONSTRUCTOR(buffer, stream_or_filename, create_buffer, explicit, _, _)
+IFCAPI_DISCOVER_CONSTRUCTOR(buffer, stream_or_filename, create_buffer_from_filename, explicit, _, _)
+IFCAPI_CONSTRUCTOR_PARAM(create_buffer_from_filename, fn, filename, const std::string&)
+IFCAPI_DISCOVER_CONSTRUCTOR(settings, ifcopenshell::geometry::Settings, create_settings, auto, _, _)
+IFCAPI_DISCOVER_CONSTRUCTOR(serializer_settings, ifcopenshell::geometry::SerializerSettings, create_serializer_settings, auto, _, _)
+IFCAPI_DISCOVER_POLICY(tree_clash_list, list_accessor, tree, clashes, tree_clash, clash_count, clash_at, "Clash index out of range")
+IFCAPI_DISCOVER_POLICY(tree_ray_intersection_list, list_accessor, tree, intersections, tree_ray_intersection, ray_intersection_count, ray_intersection_at, "Ray intersection index out of range")
+IFCAPI_DISCOVER_POLICY(settings, variant, get, set, ifcopenshell::geometry::Settings::value_variant_t, bool, bool)
+IFCAPI_DISCOVER_POLICY(settings, variant, get, set, ifcopenshell::geometry::Settings::value_variant_t, int, int, int, ifcopenshell::geometry::settings::IteratorOutputOptions, ifcopenshell::geometry::settings::FunctionStepMethod, ifcopenshell::geometry::settings::OutputDimensionalityTypes, ifcopenshell::geometry::settings::TriangulationMethod)
+IFCAPI_DISCOVER_POLICY(settings, variant, get, set, ifcopenshell::geometry::Settings::value_variant_t, double, double)
+IFCAPI_DISCOVER_POLICY(settings, variant, get, set, ifcopenshell::geometry::Settings::value_variant_t, string, std::string)
+IFCAPI_DISCOVER_POLICY(settings, variant, get, set, ifcopenshell::geometry::Settings::value_variant_t, int_set, std::set<int>)
+IFCAPI_DISCOVER_POLICY(settings, variant, get, set, ifcopenshell::geometry::Settings::value_variant_t, string_set, std::set<std::string>)
+IFCAPI_DISCOVER_POLICY(settings, variant, get, set, ifcopenshell::geometry::Settings::value_variant_t, double_list, std::vector<double>)
+IFCAPI_DISCOVER_POLICY(serializer_settings, variant, get, set, ifcopenshell::geometry::SerializerSettings::value_variant_t, bool, bool)
+IFCAPI_DISCOVER_POLICY(serializer_settings, variant, get, set, ifcopenshell::geometry::SerializerSettings::value_variant_t, int, int)
+IFCAPI_DISCOVER_POLICY(serializer_settings, variant, get, set, ifcopenshell::geometry::SerializerSettings::value_variant_t, double, double)
+IFCAPI_DISCOVER_POLICY(serializer_settings, variant, get, set, ifcopenshell::geometry::SerializerSettings::value_variant_t, string, std::string)
+IFCAPI_DISCOVER_POLICY(serializer_settings, variant, get, set, ifcopenshell::geometry::SerializerSettings::value_variant_t, int_set, std::set<int>)
+IFCAPI_DISCOVER_POLICY(triangulation, method_size, verts, verts_buffer_size)
+IFCAPI_DISCOVER_POLICY(triangulation, method_size, faces, faces_buffer_size)
+IFCAPI_DISCOVER_POLICY(triangulation, method_size, normals, normals_buffer_size)
+IFCAPI_DISCOVER_POLICY(triangulation, method_size, edges, edges_buffer_size)
+IFCAPI_DISCOVER_POLICY(triangulation, method_size, material_ids, material_ids_buffer_size)
+IFCAPI_DISCOVER_POLICY(triangulation, method_size, item_ids, item_ids_buffer_size)
+IFCAPI_DISCOVER_POLICY(triangulation, method_size, edges_item_ids, edges_item_ids_buffer_size)
+IFCAPI_DISCOVER_POLICY(triangulation, method_size, uvs, uvs_buffer_size)
+IFCAPI_DISCOVER_POLICY(triangulation, method_size, materials, material_count)
+IFCAPI_DISCOVER_POLICY(triangulation, method_at, materials, material_at, taxonomy_style, "Material index out of bounds", std::runtime_error)
+IFCAPI_DISCOVER_POLICY(taxonomy_circle, fields, inherited)
+IFCAPI_DISCOVER_POLICY(taxonomy_circle, exclude, instance, orientation, surface_style)
+IFCAPI_DISCOVER_POLICY(taxonomy_line, fields, inherited)
+IFCAPI_DISCOVER_POLICY(taxonomy_line, exclude, instance, orientation, surface_style)
+IFCAPI_DISCOVER_POLICY(taxonomy_line, as_item)
+IFCAPI_DISCOVER_POLICY(taxonomy_plane, fields, inherited)
+IFCAPI_DISCOVER_POLICY(taxonomy_plane, exclude, instance, orientation, surface_style)
+IFCAPI_DISCOVER_POLICY(taxonomy_ellipse, fields, inherited)
+IFCAPI_DISCOVER_POLICY(taxonomy_ellipse, exclude, instance, orientation, surface_style)
+IFCAPI_DISCOVER_POLICY(taxonomy_ellipse, rename, radius, radius1)
+IFCAPI_DISCOVER_POLICY(taxonomy_style, fields)
+IFCAPI_DISCOVER_POLICY(taxonomy_style, exclude, instance, orientation, surface_style, clone_, calc_hash, reverse, kind, get_color, print)
+IFCAPI_DISCOVER_POLICY(taxonomy_sphere, fields, inherited)
+IFCAPI_DISCOVER_POLICY(taxonomy_sphere, exclude, instance, orientation, surface_style)
+IFCAPI_DISCOVER_POLICY(taxonomy_torus, fields, inherited)
+IFCAPI_DISCOVER_POLICY(taxonomy_torus, exclude, instance, orientation, surface_style)
+IFCAPI_DISCOVER_POLICY(taxonomy_cylinder, fields, inherited)
+IFCAPI_DISCOVER_POLICY(taxonomy_cylinder, exclude, instance, orientation, surface_style)
+IFCAPI_DISCOVER_POLICY(taxonomy_extrusion, fields, inherited)
+IFCAPI_DISCOVER_POLICY(taxonomy_extrusion, exclude, instance, orientation, surface_style)
+IFCAPI_DISCOVER_POLICY(taxonomy_offset_curve, fields)
+IFCAPI_DISCOVER_POLICY(taxonomy_offset_curve, as_item)
+IFCAPI_DISCOVER_POLICY(taxonomy_revolve, fields, inherited)
+IFCAPI_DISCOVER_POLICY(taxonomy_revolve, exclude, instance, orientation, surface_style)
+IFCAPI_DISCOVER_POLICY(taxonomy_bspline_curve, fields)
+IFCAPI_DISCOVER_POLICY(taxonomy_bspline_curve, exclude, control_points)
+IFCAPI_DISCOVER_POLICY(taxonomy_bspline_curve, as_item)
+IFCAPI_DISCOVER_POLICY(taxonomy_bspline_curve, children, control_points, _, _, _, _, _)
+IFCAPI_DISCOVER_POLICY(taxonomy_sweep_along_curve, fields, inherited)
+IFCAPI_DISCOVER_POLICY(taxonomy_sweep_along_curve, exclude, instance, orientation, surface_style)
+IFCAPI_DISCOVER_POLICY(taxonomy_sweep_along_curve, has_fields)
+IFCAPI_DISCOVER_POLICY(taxonomy_face, fields)
+IFCAPI_DISCOVER_POLICY(taxonomy_face, as_item)
+IFCAPI_DISCOVER_POLICY(taxonomy_face, extra_field, matrix, matrix4::ptr)
+IFCAPI_DISCOVER_POLICY(taxonomy_face, children, children, _, _, _, _, _)
+IFCAPI_DISCOVER_POLICY(taxonomy_loft, fields)
+IFCAPI_DISCOVER_POLICY(taxonomy_loft, has_fields)
+IFCAPI_DISCOVER_POLICY(taxonomy_loft, field_setter, axis)
+IFCAPI_DISCOVER_POLICY(taxonomy_loft, children, children, _, _, _, add_item, ifcopenshell::geometry::taxonomy::geom_item)
+IFCAPI_DISCOVER_POLICY(taxonomy_loop, children, children, _, _, _, _, _)
+IFCAPI_DISCOVER_POLICY(taxonomy_shell, children, children, _, _, _, _, _)
+IFCAPI_DISCOVER_POLICY(taxonomy_solid, extra_field, matrix, matrix4::ptr)
+IFCAPI_DISCOVER_POLICY(taxonomy_solid, children, children, _, _, _, _, _)
+IFCAPI_DISCOVER_POLICY(taxonomy_collection, children, children, _, _, _, add_item, ifcopenshell::geometry::taxonomy::geom_item)
+IFCAPI_DISCOVER_POLICY(taxonomy_boolean_result, fields)
+IFCAPI_DISCOVER_POLICY(taxonomy_boolean_result, children, children, _, _, _, add_item, ifcopenshell::geometry::taxonomy::geom_item)
+IFCAPI_DISCOVER_POLICY(taxonomy_bspline_surface, as_item)
+IFCAPI_DISCOVER_POLICY(taxonomy_bspline_surface, array_pair, degree)
+IFCAPI_DISCOVER_POLICY(taxonomy_bspline_surface, array_pair, multiplicities)
+IFCAPI_DISCOVER_POLICY(taxonomy_bspline_surface, array_pair, knots)
+IFCAPI_DISCOVER_POLICY(tree, method_size, styles, style_count)
+IFCAPI_DISCOVER_POLICY(tree, method_at, styles, style_at, taxonomy_style, "Style index out of range", std::out_of_range)
+IFCAPI_DISCOVER_POLICY(tree, compile_guard, IFOPSH_WITH_OPENCASCADE)
+IFCAPI_DISCOVER_POLICY(tree_clash, fields)
+IFCAPI_DISCOVER_POLICY(tree_clash, rename, clash_type, type)
+IFCAPI_DISCOVER_POLICY(tree_clash, compile_guard, IFOPSH_WITH_OPENCASCADE)
+IFCAPI_DISCOVER_POLICY(tree_ray_intersection, fields)
+IFCAPI_DISCOVER_POLICY(tree_ray_intersection, exclude, instance)
+IFCAPI_DISCOVER_POLICY(tree_ray_intersection, compile_guard, IFOPSH_WITH_OPENCASCADE)
+IFCAPI_DISCOVER_POLICY(taxonomy_point3, ccomponents, get_data, "ccomponents", _)
+IFCAPI_DISCOVER_POLICY(taxonomy_direction3, ccomponents, get_data, "ccomponents", _)
+IFCAPI_DISCOVER_POLICY(taxonomy_matrix4, ccomponents, get_data, "ccomponents", _)
+IFCAPI_DISCOVER_POLICY(taxonomy_colour, ccomponents, get_data, "ccomponents", _)
+IFCAPI_DISCOVER_POLICY(transformation, ccomponents, matrix, "data()->ccomponents", _)
+
+IFCAPI_DISCOVER_FUNCTION(ifcopenshell::geometry, helmert_curve_point)
 
 namespace ifcgeom::bindings {
