+++ src/ifcwrap/binding_generator/specs/cpp/ifcparse.hpp
@@ -49,6 +49,137 @@ IFCAPI_HANDLE(inverse_attribute, ifcopenshell::inverse_attribute, none)
 IFCAPI_HANDLE(attribute_value, attribute_value, none, value) struct ifcopenshell_parse_attribute_value_t;
 IFCAPI_HANDLE(instance_list, std::vector<express::Base>, none, value) struct ifcopenshell_parse_instance_list_t;
 
+// Native methods included in the stable C ABI.
+IFCAPI_DISCOVER_METHOD(file, create, create, const ifcopenshell::declaration*, int)
+IFCAPI_DISCOVER_METHOD(file, get_inverse, get_inverses_by_declaration, int, const ifcopenshell::declaration*, int)
+IFCAPI_DISCOVER_METHOD(file, instances_by_type, by_type, const std::string&)
+IFCAPI_DISCOVER_METHOD(file, instances_by_type_excl_subtypes, by_type_excl_subtypes, const std::string&)
+IFCAPI_DISCOVER_METHOD(file, add_entity, add, const express::Base&, int)
+IFCAPI_DISCOVER_METHOD(file, add_type_ref, add_type_ref, const express::Base&)
+IFCAPI_DISCOVER_METHOD(file, batch, batch)
+IFCAPI_DISCOVER_METHOD(file, build_inverses, build_inverses)
+IFCAPI_DISCOVER_METHOD(file, build_inverses_, build_inverses_, const express::Base&)
+IFCAPI_DISCOVER_METHOD(file, bypass_type, bypass_type, const std::string&)
+IFCAPI_DISCOVER_METHOD(file, create_timestamp, create_timestamp)
+IFCAPI_DISCOVER_METHOD(file, fresh_id, fresh_id)
+IFCAPI_DISCOVER_METHOD(file, get_inverse_indices_by_id, get_inverse_indices_by_id, int)
+IFCAPI_DISCOVER_METHOD(file, get_max_id, get_max_id)
+IFCAPI_DISCOVER_METHOD(file, get_total_inverses, get_total_inverses_by_id, int)
+IFCAPI_DISCOVER_METHOD(file, ifcroot_type, ifcroot_type)
+IFCAPI_DISCOVER_METHOD(file, initialize, initialize, const std::string&, ifcopenshell::filetype, bool)
+IFCAPI_DISCOVER_METHOD(file, instance_by_guid, by_guid, const std::string&)
+IFCAPI_DISCOVER_METHOD(file, instance_by_id, by_id, int)
+IFCAPI_DISCOVER_METHOD(file, instances_by_reference, instances_by_reference, int)
+IFCAPI_DISCOVER_METHOD(file, process_deletion_inverse, process_deletion_inverse, const express::Base&)
+IFCAPI_DISCOVER_METHOD(file, recalculate_id_counter, recalculate_id_counter)
+IFCAPI_DISCOVER_METHOD(file, remove_entity, remove, const express::Base&)
+IFCAPI_DISCOVER_METHOD(file, remove_type_ref, remove_type_ref, const express::Base&)
+IFCAPI_DISCOVER_METHOD(file, reset_identity_cache, reset_identity_cache)
+IFCAPI_DISCOVER_METHOD(file, schema, schema)
+IFCAPI_DISCOVER_METHOD(file, traverse, traverse, const express::Base&, int)
+IFCAPI_DISCOVER_METHOD(file, traverse_breadth_first, traverse_breadth_first, const express::Base&, int)
+IFCAPI_DISCOVER_METHOD(file, unbatch, unbatch)
+IFCAPI_DISCOVER_METHOD(instance, declaration, declaration)
+IFCAPI_DISCOVER_METHOD(instance, file, file)
+IFCAPI_DISCOVER_METHOD(instance, get_attribute_value, get_argument, unsigned long)
+IFCAPI_DISCOVER_METHOD(instance, id, id)
+IFCAPI_DISCOVER_METHOD(instance, identity, identity)
+IFCAPI_DISCOVER_METHOD(schema, declaration_by_name, declaration_by_name, const std::string&)
+IFCAPI_DISCOVER_METHOD(schema, declaration_by_name, declaration_by_index, unsigned long)
+IFCAPI_DISCOVER_METHOD(schema, declarations, declarations)
+IFCAPI_DISCOVER_METHOD(schema, entities, entities)
+IFCAPI_DISCOVER_METHOD(schema, enumeration_types, enumeration_types)
+IFCAPI_DISCOVER_METHOD(schema, name, name)
+IFCAPI_DISCOVER_METHOD(schema, select_types, select_types)
+IFCAPI_DISCOVER_METHOD(schema, type_declarations, type_declarations)
+IFCAPI_DISCOVER_METHOD(declaration, is, is_a, const std::string&)
+IFCAPI_DISCOVER_METHOD(declaration, as_entity, as_entity)
+IFCAPI_DISCOVER_METHOD(declaration, as_enumeration_type, as_enumeration_type)
+IFCAPI_DISCOVER_METHOD(declaration, as_select_type, as_select_type)
+IFCAPI_DISCOVER_METHOD(declaration, as_type_declaration, as_type_declaration)
+IFCAPI_DISCOVER_METHOD(declaration, index_in_schema, index_in_schema)
+IFCAPI_DISCOVER_METHOD(declaration, name, name)
+IFCAPI_DISCOVER_METHOD(declaration, name_uc, name_uc)
+IFCAPI_DISCOVER_METHOD(declaration, schema, schema)
+IFCAPI_DISCOVER_METHOD(declaration, type, type)
+IFCAPI_DISCOVER_METHOD(type_declaration, as_type_declaration, as_type_declaration)
+IFCAPI_DISCOVER_METHOD(type_declaration, declared_type, declared_type)
+IFCAPI_DISCOVER_METHOD(select_type, as_select_type, as_select_type)
+IFCAPI_DISCOVER_METHOD(select_type, select_list, select_list)
+IFCAPI_DISCOVER_METHOD(enumeration, as_enumeration_type, as_enumeration_type)
+IFCAPI_DISCOVER_METHOD(enumeration, enumeration_items, enumeration_items)
+IFCAPI_DISCOVER_METHOD(enumeration, lookup_enum_offset, lookup_enum_offset, const std::string&)
+IFCAPI_DISCOVER_METHOD(enumeration, lookup_enum_value, lookup_enum_value, unsigned long)
+IFCAPI_DISCOVER_METHOD(parameter_type, as_aggregation_type, as_aggregation_type)
+IFCAPI_DISCOVER_METHOD(parameter_type, as_named_type, as_named_type)
+IFCAPI_DISCOVER_METHOD(parameter_type, as_simple_type, as_simple_type)
+IFCAPI_DISCOVER_METHOD(named_type, is, is_a, const std::string&)
+IFCAPI_DISCOVER_METHOD(named_type, as_named_type, as_named_type)
+IFCAPI_DISCOVER_METHOD(named_type, declared_type, declared_type)
+IFCAPI_DISCOVER_METHOD(simple_type, as_simple_type, as_simple_type)
+IFCAPI_DISCOVER_METHOD(simple_type, declared_type, declared_type)
+IFCAPI_DISCOVER_METHOD(aggregation_type, as_aggregation_type, as_aggregation_type)
+IFCAPI_DISCOVER_METHOD(aggregation_type, bound1, bound1)
+IFCAPI_DISCOVER_METHOD(aggregation_type, bound2, bound2)
+IFCAPI_DISCOVER_METHOD(aggregation_type, type_of_element, type_of_element)
+IFCAPI_DISCOVER_METHOD(header, owner_file, file)
+IFCAPI_DISCOVER_METHOD(header, file_description, file_description)
+IFCAPI_DISCOVER_METHOD(header, file_name, file_name)
+IFCAPI_DISCOVER_METHOD(header, file_schema, file_schema)
+IFCAPI_DISCOVER_METHOD(file_description, Class, class)
+IFCAPI_DISCOVER_METHOD(file_description, description, description)
+IFCAPI_DISCOVER_METHOD(file_description, implementation_level, implementation_level)
+IFCAPI_DISCOVER_METHOD(file_description, initialize, initialize, std::vector<std::string>, std::string)
+IFCAPI_DISCOVER_METHOD(file_description, setdescription, setdescription, const std::vector<std::string>&)
+IFCAPI_DISCOVER_METHOD(file_description, setimplementation_level, setimplementation_level, const std::string&)
+IFCAPI_DISCOVER_METHOD(file_name, Class, class)
+IFCAPI_DISCOVER_METHOD(file_name, author, author)
+IFCAPI_DISCOVER_METHOD(file_name, authorization, authorization)
+IFCAPI_DISCOVER_METHOD(file_name, initialize, initialize, std::string, std::string, std::vector<std::string>, std::vector<std::string>, std::string, std::string, std::string)
+IFCAPI_DISCOVER_METHOD(file_name, name, name)
+IFCAPI_DISCOVER_METHOD(file_name, organization, organization)
+IFCAPI_DISCOVER_METHOD(file_name, originating_system, originating_system)
+IFCAPI_DISCOVER_METHOD(file_name, preprocessor_version, preprocessor_version)
+IFCAPI_DISCOVER_METHOD(file_name, setauthor, setauthor, const std::vector<std::string>&)
+IFCAPI_DISCOVER_METHOD(file_name, setauthorization, setauthorization, const std::string&)
+IFCAPI_DISCOVER_METHOD(file_name, setname, setname, const std::string&)
+IFCAPI_DISCOVER_METHOD(file_name, setorganization, setorganization, const std::vector<std::string>&)
+IFCAPI_DISCOVER_METHOD(file_name, setoriginating_system, setoriginating_system, const std::string&)
+IFCAPI_DISCOVER_METHOD(file_name, setpreprocessor_version, setpreprocessor_version, const std::string&)
+IFCAPI_DISCOVER_METHOD(file_name, settime_stamp, settime_stamp, const std::string&)
+IFCAPI_DISCOVER_METHOD(file_name, time_stamp, time_stamp)
+IFCAPI_DISCOVER_METHOD(file_schema, Class, class)
+IFCAPI_DISCOVER_METHOD(file_schema, initialize, initialize, std::vector<std::string>)
+IFCAPI_DISCOVER_METHOD(file_schema, schema_identifiers, schema_identifiers)
+IFCAPI_DISCOVER_METHOD(file_schema, setschema_identifiers, setschema_identifiers, const std::vector<std::string>&)
+IFCAPI_DISCOVER_METHOD(entity, attribute_index, attribute_index, const std::string&)
+IFCAPI_DISCOVER_METHOD(entity, all_attributes, all_attributes)
+IFCAPI_DISCOVER_METHOD(entity, all_inverse_attributes, all_inverse_attributes)
+IFCAPI_DISCOVER_METHOD(entity, as_entity, as_entity)
+IFCAPI_DISCOVER_METHOD(entity, attribute_by_index, attribute_by_index, unsigned long)
+IFCAPI_DISCOVER_METHOD(entity, attribute_count, attribute_count)
+IFCAPI_DISCOVER_METHOD(entity, attributes, attributes)
+IFCAPI_DISCOVER_METHOD(entity, derived, derived)
+IFCAPI_DISCOVER_METHOD(entity, inverse_attributes, inverse_attributes)
+IFCAPI_DISCOVER_METHOD(entity, is_abstract, is_abstract)
+IFCAPI_DISCOVER_METHOD(entity, set_attributes, set_attributes, const std::vector<const ifcopenshell::attribute*>&, const std::vector<bool>&)
+IFCAPI_DISCOVER_METHOD(entity, set_inverse_attributes, set_inverse_attributes, const std::vector<const ifcopenshell::inverse_attribute*>&)
+IFCAPI_DISCOVER_METHOD(entity, set_subtypes, set_subtypes, const std::vector<const ifcopenshell::entity*>&)
+IFCAPI_DISCOVER_METHOD(entity, subtypes, subtypes)
+IFCAPI_DISCOVER_METHOD(entity, supertype, supertype)
+IFCAPI_DISCOVER_METHOD(attribute, name, name)
+IFCAPI_DISCOVER_METHOD(attribute, optional, optional)
+IFCAPI_DISCOVER_METHOD(attribute, type_of_attribute, type_of_attribute)
+IFCAPI_DISCOVER_METHOD(inverse_attribute, attribute_reference, attribute_reference)
+IFCAPI_DISCOVER_METHOD(inverse_attribute, bound1, bound1)
+IFCAPI_DISCOVER_METHOD(inverse_attribute, bound2, bound2)
+IFCAPI_DISCOVER_METHOD(inverse_attribute, entity_reference, entity_reference)
+IFCAPI_DISCOVER_METHOD(inverse_attribute, name, name)
+IFCAPI_DISCOVER_METHOD(instance_streamer, bypassed_instances, bypassed_instances)
+IFCAPI_DISCOVER_METHOD(instance_streamer, has_semicolon, has_semicolon)
+IFCAPI_DISCOVER_METHOD(instance_streamer, push_page, push_page, const std::string&)
+IFCAPI_DISCOVER_METHOD(instance_streamer, semicolon_count, semicolon_count)
+
 namespace ifcopenshell::capi {
 
 void set_feature(const std::string& name, bool value);
