+++ src/ifcwrap/binding_generator/specs/cpp/spec_macros.h
@@ -8,6 +8,12 @@
 #define IFCAPI_COPY
 #define IFCAPI_STATIC
 #define IFCAPI_HANDLE(cpp_type, destructor, ...)
+#define IFCAPI_DISCOVER_METHOD(handle, cpp_name, expose_as, ...)
+#define IFCAPI_DISCOVER_FUNCTION(cpp_namespace, name)
+#define IFCAPI_DISCOVER_CONSTRUCTOR(handle, cpp_class, expose_as, mode, guard, message)
+#define IFCAPI_CONSTRUCTOR_PARAM(expose_as, source_name, public_name, cpp_type)
+#define IFCAPI_DISCOVER_POLICY(handle, operation, ...)
+#define IFCAPI_INT32(cpp_type) cpp_type
 #define IFCAPI_RESULT_STRUCT(...)
 
 #endif // IFCWRAP_BINDING_GENERATOR_SPEC_MACROS_H
